Smoothstep to double

+++ EventBaseMod/src/main/java/suso/event_base/util/MiscUtil.java
@@ -3,8 +3,8 @@ package suso.event_base.util;
 import net.minecraft.util.math.MathHelper;
 
 public class MiscUtil {
-    public static float smoothStep(float from, float to, float x) {
-        x = MathHelper.clamp((x - from) / (to - from), 0.0f, 1.0f);
-        return x * x * x * (x * (x * 6.0f - 15.0f) + 10.0f);
+    public static double smoothStep(double from, double to, double x) {
+        x = MathHelper.clamp((x - from) / (to - from), 0.0, 1.0);
+        return x * x * x * (x * (x * 6.0 - 15.0) + 10.0);
     }
 }
